requirements.txt
aiogram==3.4.1
telethon==1.34.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
cryptobot-python==0.4.1
